[#91]:amelia: fix four bugs in append and buffered transfer mechanics
- Move h5::impl::free + unique_ptr alias to H5misc.hpp (earlier in core
  include chain) so H5Zpipeline.hpp can use the correct deleter
- H5Zpipeline: change ptr0/ptr1 from std::unique_ptr<char> (calls delete)
  to h5::impl::unique_ptr<char> (calls std::free), fixing UB on aligned_alloc memory;
  switch set_cache allocation from move-construct to .reset()
- H5Dappend: fix move assignment operator returning pt instead of *this
- H5Dappend: fix flush() extent formula — replace no-op
  (*current_dims += *current_dims % *chunk_dims) and broken row-count
  arithmetic with a single *current_dims += *chunk_dims (mirrors scalar
  append path, required for write_chunk to stay in-bounds)
- H5Dappend: add const char* append specialization so C-string literals
  route to the VL string path instead of falling through to the raw-buffer
  pointer overload
